Add cml service
